Update Assignment EX7.c

+++ C Language/C Basics Assignment/Assignment EX7.c
@@ -8,20 +8,20 @@
 #include <stdlib.h>
 
 int main() {
-	int a,b;
+	float a,b;
 	printf("Enter value of a: ");
 	fflush(stdin); fflush(stdout);
-	scanf("%d",&a);
+	scanf("%f",&a);
 	printf("Enter value of b: ");
 	fflush(stdin); fflush(stdout);
-	scanf("%d",&b);
+	scanf("%f",&b);
 	fflush(stdin); fflush(stdout);
 
 	a=a+b;
 	b=a-b;
 	a=a-b;
-	printf("After swapping, value of a = %d \n",a);
-	printf("After swapping, value of b = %d \n",b);
+	printf("After swapping, value of a = %f \n",a);
+	printf("After swapping, value of b = %f \n",b);
 	fflush(stdin); fflush(stdout);
 	return 0;
 }
